Update MemberDiv for new site layout
Phone, fax, party and contact information are no longer listed in
individual member rows. Contact form urls are no longer listed in either
member rows or on member pages.

Phone, fax and party info will be captured in MemberPage which will be added
in a forthcoming commit.
