+++ scraper.rb
@@ -10,43 +10,8 @@ require 'scraperwiki'
 # OpenURI::Cache.cache_path = '.cache'
 require 'scraped_page_archive/open-uri'
 
-class MemberDiv < Scraped::HTML
-  field :id do
-    noko.css('a.more-info/@href').text[/rep=(\d+)/, 1]
-  end
+require_rel 'lib'
 
-  field :name do
-    noko.xpath('.//span[@class="info"]//span[@class="name"]/text()').text.split(' - ').first.tidy.sub('Hon. ', '')
-  end
-
-  field :party do
-    noko.css('.info .party').text.tidy
-  end
-
-  field :area do
-    noko.css('.info .district').text.tidy
-  end
-
-  field :image do
-    noko.css('.identity img/@src').text
-  end
-
-  field :phone do
-    noko.xpath('.//span[@class="data-type" and contains(.,"Tel:")]').map { |n| n.text.sub('Tel:', '').tidy }.reject(&:empty?).join(' / ')
-  end
-
-  field :fax do
-    noko.xpath('.//span[@class="data-type" and contains(.,"Fax:")]').map { |n| n.text.sub('Fax:', '').tidy }.reject(&:empty?).join(' / ')
-  end
-
-  field :contact_form do
-    noko.css('a.mail/@href').text
-  end
-
-  field :source do
-    noko.css('a.more-info/@href').text
-  end
-end
 
 class MembersPage < Scraped::HTML
   decorator Scraped::Response::Decorator::CleanUrls
